+++ texlite/utils.py
@@ -29,3 +29,12 @@ def read_json(path):
             return json.load(f)
     except FileNotFoundError:
         msg.error(f'JSON file "{path}" could not be found', halt=True)
+
+
+def read_file_as_list(path):
+
+    try:
+        with open(Path(path)) as f:
+            return [line.strip('\n') for line in f.readlines()]
+    except FileNotFoundError:
+        msg.error(f'File "{path}" could not be found', halt=True)